backend/requirements.txt
langchain==0.1.12
langchain-openai==0.0.8
# langchain-community==0.0.27
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.3
python-multipart==0.0.9
faiss-cpu==1.7.4 